feat: bump pb to v3

+++ cmd/cp-main.go
@@ -498,10 +498,10 @@ loop:
 				if progressReader, pgok := pg.(*progressBar); pgok {
 					if progressReader.Get() > 0 {
 						writeContSize := (int)(cpURLs.SourceContent.Size)
-						totalPGSize := (int)(progressReader.Total)
+						totalPGSize := (int)(progressReader.Total())
 						written := (int)(progressReader.Get())
 						if totalPGSize > writeContSize && written > writeContSize {
-							progressReader.Set((written - writeContSize))
+							progressReader.Set(int64(written - writeContSize))
 							progressReader.Update()
 						}
 					}
